Add advancement and attack tables, and character dashboard features
This commit introduces the criterion for character advancement and attack by creating advancement and attack tables. On the UI, a Character Dashboard is added to show progress/skill of characters. In addition to these, this commit includes the implementation of several character abilities and the required seeders, factories and requests. New dependencies are added to the package.json file and composer.json file.